fix: extract duplicated extractJson, add TODO for parseUsage (PR #45)
