+++ src/core/agents/junie.ts
@@ -4,17 +4,7 @@ import {
   type TokenUsage,
 } from "./types.js";
 import { TextBasedAgent, type TextBasedAgentDeps } from "./text-based-agent.js";
-
-const JSON_FENCE_RE = /```(?:json)?\s*\n([\s\S]*?)\n\s*```/;
-const TRAILING_JSON_RE = /\{[\s\S]*\}\s*$/;
-
-function extractJson(text: string): string {
-  const fenceMatch = text.match(JSON_FENCE_RE);
-  if (fenceMatch) return fenceMatch[1];
-  const trailingMatch = text.match(TRAILING_JSON_RE);
-  if (trailingMatch) return trailingMatch[0];
-  return text;
-}
+import { extractJson } from "./json-utils.js";
 
 export class JunieAgent extends TextBasedAgent {
   name = "junie";
@@ -47,6 +37,7 @@ export class JunieAgent extends TextBasedAgent {
   }
 
   protected parseUsage(_stdout: string): TokenUsage {
+    // TODO: Extract token usage from Junie CLI output if available
     return {
       inputTokens: 0,
       outputTokens: 0,
